streamlit_app/requirements.txt
streamlit
opencv-python
numpy
torch
torchvision
Pillow
ultralytics
